+++ src/pages/Detail/css/Detail.module.css
