+++ src/pages/Detail/css/History.module.css
